+++ src/main/kotlin/Main.kt
@@ -3,6 +3,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
+import ui.App
 
 fun main() = application {
     val state = rememberWindowState(size = DpSize(1040.dp, 680.dp))